src/requirements.txt
flask
flask-sqlalchemy
python-dotenv
flask_login 
flask_bcrypt 
flask_wtf 
wtforms 
email_validator
pyyaml